+++ lib/auth/logout_page.dart
@@ -1,13 +1,15 @@
-import 'package:capi_stonsker/nav/bottom_nav_bar.dart';
+/*
+ * This is quite literally just a page to log out, for testing purposes.
+ *
+ * TODO This provides basic functionality and will be improved in the beta release
+ */
+
+import 'package:capi_stonsker/app_nav/bottom_nav_bar.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'log_in_page.dart';
-import 'side_menu.dart';
-import 'sign_up_page.dart';
-
-//TODO this is quite literally just a page to log out, for testing purposes.
-// It is basic functionality. It will be improved in the beta release
+import '../app_nav/side_menu.dart';
 
 class LogoutPage extends StatelessWidget {
   LogoutPage({Key? key}) : super(key: key);
@@ -44,7 +46,6 @@ class LogoutPage extends StatelessWidget {
                   )
               );
             }
-
           ),
         ),
       drawer: SideMenu(),
@@ -52,8 +53,6 @@ class LogoutPage extends StatelessWidget {
     );
   }
 }
-
-
 
 class LoginSignupButton extends StatelessWidget {
   final String title;
