+++ lib/markers/marker_list_page.dart
@@ -1,8 +1,18 @@
-import 'package:capi_stonsker/Widgets/side_menu.dart';
-import 'package:capi_stonsker/nav/bottom_nav_bar.dart';
+/*
+ * This page is accessible from the home page and provides a list view
+ * of the markers, sorted by distance from the user.
+ *
+ * A marker can be navigated to by pressing the nav button
+ * to the right of the marker's name
+ *
+ * Tapping a marker opens that marker's full info page
+ */
+
+import 'package:capi_stonsker/app_nav/side_menu.dart';
+import 'package:capi_stonsker/app_nav/bottom_nav_bar.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import '../src/locations.dart' as locs;
+import 'package:capi_stonsker/markers/locations.dart' as locs;
 
 class MarkerListPage extends StatefulWidget {
   MarkerListPage({Key? key}) : super(key: key);
