+++ lib/routing/plan_route_page.dart
@@ -1,10 +1,19 @@
-import 'package:capi_stonsker/nav/bottom_nav_bar.dart';
+/*
+ * The idea of this page is to plan complex routes, such as routes that
+ * visit multiple different markers
+ *
+ * This page is not yet implemented as of the Proof of Concept release
+ *
+ * Consider: we may end up removing this page and not implementing this feature
+ */
+
+import 'package:capi_stonsker/app_nav/bottom_nav_bar.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'side_menu.dart';
+import 'package:capi_stonsker/app_nav/side_menu.dart';
 
-class FriendsPage extends StatelessWidget {
-  FriendsPage({Key? key}) : super(key: key);
+class PlanRoutePage extends StatelessWidget {
+  PlanRoutePage({Key? key}) : super(key: key);
   GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey();
 
   @override
@@ -20,7 +29,7 @@ class FriendsPage extends StatelessWidget {
               Navigator.of(context).pop();
             }
         ),
-        title: Text("Friends Page"),
+        title: Text("Plan Route Page"),
         backgroundColor: Colors.blueGrey,
       ),
       drawer: SideMenu(),
@@ -28,3 +37,6 @@ class FriendsPage extends StatelessWidget {
     );
   }
 }
+
+
+
